scripts/requirements.txt
gtts==2.5.1
httpx==0.27.0
mediapipe==0.10.11
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
paho_mqtt==2.1.0
playaudio==1.0.5
psutil==5.9.8
transitions==0.9.1
